粘豆包/api/zhandoubao/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiomysql==0.2.0
PyJWT==2.9.0
cryptography==43.0.0
openai==1.50.0
pydantic==2.9.0
python-dotenv==1.0.1
sse-starlette==2.1.0
httpx==0.27.0
